Fix of a segmentation fault on long labels
The program 'chsmack' was crashing with a 'segmentation fault'
when called with a label longer than 255 characters, as, for example:

 chsmack -a $(head -c256 /dev/zero | tr '\0' x) afile

The problem came from the function getopt_long and from how it was used.
Internal documentation of getopt_long states that:

 LONGIND (the 5th parameter) returns the index in LONGOPT of the
 long-named option found. It is only valid when a long-named option
 has been found by the most recent call.

Removing any reference to its 5th parameter (replaced with NULL) solves
the problem.