chsmack: using linux constants for Smack's names
It is better to rely on the centralisation of names in a
single file: it can avoid errors.

The declarations of <linux/xattr.h> contains the definitions
of the names of the Smack's security attributes.

As Smack only runs on Linux and libsmack doesn't provides
a centralized version of the names of the security attributes
(because it were not needed), using the linux header seems
to be a good choice.

Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>