chsmack: add dereference option
Add the options that allow to follow the symbolic
links instead of modifying it. There is no really
need but it may help.

The text of the short option is moved.
It prepares to scan options in two passes. It also
centralize the definition of options, avoiding to have
many lines between the definitions.

Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>