libsmack: add functions for setting and removing labels on files
Jóse Bollo implemented two functions as part of this various
improvements for the chsmack command-line utility:

- smack_set_label_for_path() (see f1dfd85)
- smack_remove_label_for_path() (see 5da1a22)

Since they are generally useful, they should be part of the
API in libsmack 1.1.

This patch migrates these functions to libsmack and exports
the symbols. Also, the chsmack is modified to use the new API
instead of the internal functions.

[jsakkine: rewrote the patch description]

Signed-off-by: Rafal Krypa <r.krypa@samsung.com>
Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>