libsmack: Terminate attribute string
The smack_new_label_from_path function reads an xattr
that may not be null byte terminated. This occurs in the
SMACK64TRANSMUTE case. Technically, the transmute attribute
isn't a label, so this function shouldn't be used to fetch
the value, but we'll let that go. This is just good string
hygiene in any case.

This is an issue because chsmack prints transmute="TRUE0"
without this fix.

Signed-off-by: Casey Schaufler <casey.schaufler@intel.com>